Renamed RFBeamLoadingCavityPass and exclude from windows, limit complex to RFBeamLoadingCavityPass, exc
